requirements.txt
exchange_calendars==4.11.1
iniconfig==2.1.0
korean-lunar-calendar==0.3.1
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pandas_market_calendars==5.1.1
pluggy==1.6.0
Pygments==2.19.2
pyluach==2.2.0
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
structlog==25.4.0
toolz==1.0.0
tzdata==2025.2
websockets==15.0.1
redis>=5
python-dotenv